requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-Babel==4.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.1.8
blinker==1.9.0
itsdangerous==2.2.0
MarkupSafe==3.0.2

# Internationalization
Babel==2.17.0
pytz==2025.2

# Database - Supabase and dependencies
supabase==2.0.0
postgrest==0.13.2
gotrue==1.3.1
realtime==1.0.6
storage3==0.6.1
supafunc==0.3.3

# HTTP clients
httpx==0.24.1
httpcore==0.17.3
h11==0.14.0
certifi==2025.10.5
anyio==4.11.0
sniffio==1.3.1

# Data validation
pydantic==2.12.4
pydantic_core==2.41.5
annotated-types==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0

# Utilities
python-dotenv==1.2.0
python-dateutil==2.9.0.post0
Deprecated==1.2.18
deprecation==2.1.0
wrapt==1.17.2
packaging==24.2
StrEnum==0.4.15
six==1.17.0
idna==3.11

# IoT & Hardware
paho-mqtt==1.6.1
pyserial==3.5
websockets==12.0

# PDF & Image processing
pikepdf==9.5.2
pillow==11.1.0
lxml==5.3.1

